create managers module and move all managers there (#419)
