--- pytripgui/canvas_vc/plotter/__init__.py
@@ -1,7 +1,0 @@
-from pytripgui.canvas_vc.plotter.coordinate_info import CoordinateInfo
-from pytripgui.canvas_vc.plotter.plotting_manager import PlottingManager
-
-# from https://docs.python.org/3/tutorial/modules.html
-# if a package's __init__.py code defines a list named __all__,
-# it is taken to be the list of module names that should be imported when from package import * is encountered.
-__all__ = ['CoordinateInfo', 'PlottingManager']
