+++ pytripgui/canvas_vc/plotter/bars/__init__.py
@@ -1,3 +1,6 @@
+from pytripgui.canvas_vc.plotter.bars.ctx_bar import CtxBar
+from pytripgui.canvas_vc.plotter.bars.dos_bar import DosBar
+from pytripgui.canvas_vc.plotter.bars.let_bar import LetBar
 from pytripgui.canvas_vc.plotter.bars.projection_enum import BarProjection
 
 # from https://docs.python.org/3/tutorial/modules.html
