+++ pytripgui/canvas_vc/plotter/canvas_plotter.py
@@ -3,8 +3,7 @@ from PyQt5.QtWidgets import QSizePolicy
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.figure import Figure
 
-from pytripgui.canvas_vc.plotter.blit_manager import BlitManager
-from pytripgui.canvas_vc.plotter import PlottingManager
+from pytripgui.canvas_vc.plotter.managers import PlottingManager, BlitManager
 
 
 class CanvasPlotter(FigureCanvas):
