+++ pytripgui/canvas_vc/plotter/managers/__init__.py
@@ -1,0 +1,5 @@
+from pytripgui.canvas_vc.plotter.managers.blit_manager import BlitManager
+from pytripgui.canvas_vc.plotter.managers.placement_manager import PlacementManager
+from pytripgui.canvas_vc.plotter.managers.plotting_manager import PlottingManager
+
+__all__ = ['BlitManager', 'PlacementManager', 'PlottingManager']
