+++ pytripgui/canvas_vc/plotter/managers/blit_manager.py
