+++ pytripgui/canvas_vc/plotter/managers/placement_manager.py
